backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
daphne==4.1.2
defusedxml==0.8.0rc2
Django==4.2.14
django-redis==5.4.0
django-render-block==0.9.2
django-templated-email==3.0.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
install==1.3.5
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.0.8
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python3-openid==3.2.0
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
zope.interface==6.4.post2
